Refactor structure of folder.
