+++ maven/IMG2PDF/src/main/java/org/vincentyeh/IMG2PDF/commandline/action/ConvertAction.java
@@ -13,7 +13,7 @@ import org.vincentyeh.IMG2PDF.SharedSpace;
 import org.vincentyeh.IMG2PDF.commandline.parser.core.HandledException;
 import org.vincentyeh.IMG2PDF.converter.PDFConverter;
 import org.vincentyeh.IMG2PDF.task.Task;
-import org.vincentyeh.IMG2PDF.task.converter.TaskListConverter;
+import org.vincentyeh.IMG2PDF.task.TaskListConverter;
 import org.vincentyeh.IMG2PDF.util.file.FileUtils;
 
 public class ConvertAction implements Action {
